feat: backend completed

+++ backend/src/watchlist/service/watchlist.service.ts
@@ -1,5 +1,6 @@
 import { ApiResponse } from "../../common/types/response";
 import { WatchlistRepository, watchlistRepo } from "../repository/watchlist.repository";
+import { WatchlistEntityType } from "../types/WatchlistEntityType";
 
 export class WatchListService {
   protected watchlistRepo: WatchlistRepository;
@@ -8,12 +9,27 @@ export class WatchListService {
     this.watchlistRepo = watchlistRepo;
   }
 
-  async add(userId: number, stockId: number): Promise<void> {
+  async add(userId: number, stockId: number): Promise<ApiResponse<string>> {
+    const exists = await this.watchlistRepo.findBySymbol(userId, stockId);
+    if (exists.length > 0)
+      return { success: false, error: "Stock already exists in watchlist" };
+
     await this.watchlistRepo.create({ user_id: userId, stock_id: stockId });
+
+    return { success: true, data: "Successfully added to watchlist" };
   }
 
-  async remove(userId: number, stockId: number): Promise<void> {
+  async remove(userId: number, stockId: number): Promise<ApiResponse<string>> {
+    const exists = await this.watchlistRepo.findBySymbol(userId, stockId);
+    if (exists.length == 0)
+      return { success: false, error: "Stock does not exists in watchlist" };
+
     await this.watchlistRepo.deleteBySymbol(userId, stockId);
+    return { success: true, data: "Successfully removed from watchlist" };
+  }
+
+  async findAllByUserId(userId: number): Promise<WatchlistEntityType[]> {
+    return await this.watchlistRepo.findAllByUserId(userId);
   }
 }
